pyproject.toml
[project]
name = "money-printer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.156.0",
    "praw>=7.8.1",
    "python-dotenv>=1.0.1",
    "pytubefix>=8.8.4",
]
